requirements.txt
Package                 Version
----------------------- ------------------
absl-py                 2.1.0
addict                  2.4.0
affine                  2.4.0
aiohttp                 3.9.5
aiosignal               1.3.1
albumentations          1.0.3
anaconda-anon-usage     0.4.4
annotated-types         0.7.0
antlr4-python3-runtime  4.9.3
archspec                0.2.3
asttokens               2.4.1
async-timeout           4.0.3
attrs                   23.2.0
bleach                  6.1.0
blis                    0.7.11
boltons                 23.0.0
Brotli                  1.0.9
catalogue               2.0.10
certifi                 2024.6.2
cffi                    1.16.0
charset-normalizer      2.0.4
click                   8.1.7
click-plugins           1.1.1
cligj                   0.7.2
cloudpathlib            0.18.1
conda                   24.5.0
conda-content-trust     0.2.0
conda-libmamba-solver   24.1.0
conda-package-handling  2.3.0
conda_package_streaming 0.10.0
confection              0.1.5
contourpy               1.2.1
cryptography            42.0.5
cycler                  0.12.1
cymem                   2.0.8
Cython                  3.0.10
dataclasses             0.6
decorator               5.1.1
distro                  1.9.0
docker-pycreds          0.4.0
effdet                  0.2.4
exceptiongroup          1.2.1
executing               2.0.1
fastai                  2.5.6
fastcore                1.3.29
fastdownload            0.0.7
fastprogress            1.0.3
filelock                3.15.4
fire                    0.6.0
fonttools               4.53.0
frozendict              2.4.2
frozenlist              1.4.1
fsspec                  2024.6.1
fvcore                  0.1.5.post20221221
gitdb                   4.0.11
GitPython               3.1.43
grpcio                  1.64.1
huggingface-hub         0.23.4
icedata                 0.5.1
icevision               0.12.0
idna                    3.7
imageio                 2.34.2
importlib_metadata      8.0.0
importlib_resources     6.4.0
iopath                  0.1.10
ipykernel               5.5.6
ipython                 8.18.1
ipython-genutils        0.2.0
jedi                    0.19.1
Jinja2                  3.1.4
joblib                  1.4.2
jsonpatch               1.33
jsonpointer             2.1
jupyter_client          8.6.2
jupyter_core            5.7.2
kaggle                  1.6.14
kiwisolver              1.4.5
langcodes               3.4.0
language_data           1.2.0
libmambapy              1.5.8
lightning-utilities     0.11.3.post0
loguru                  0.7.2
marisa-trie             1.2.0
Markdown                3.6
markdown-it-py          3.0.0
MarkupSafe              2.1.5
matplotlib              3.9.0
matplotlib-inline       0.1.7
mdurl                   0.1.2
menuinst                2.1.1
mmcv-full               1.3.17
mmdet                   2.17.0
mmsegmentation          0.20.2
multidict               6.0.5
murmurhash              1.0.10
networkx                3.2.1
nose                    1.3.7
numpy                   1.26.4
omegaconf               2.3.0
opencv-python           4.8.1.78
opencv-python-headless  4.10.0.84
packaging               23.2
pandas                  2.2.2
parso                   0.8.4
pexpect                 4.9.0
Pillow                  8.4.0
pip                     24.0
platformdirs            3.10.0
pluggy                  1.0.0
portalocker             2.10.0
preshed                 3.0.9
prettytable             3.10.0
prompt_toolkit          3.0.47
protobuf                4.25.3
psutil                  6.0.0
ptyprocess              0.7.0
pure-eval               0.2.2
pybboxes                0.1.6
pycocotools             2.0.8
pycosat                 0.6.6
pycparser               2.21
pydantic                2.7.4
pydantic_core           2.18.4
Pygments                2.18.0
pyparsing               3.1.2
PySocks                 1.7.1
python-dateutil         2.9.0.post0
python-slugify          8.0.4
pytorch-lightning       1.9.5
pytz                    2024.1
PyWavelets              1.6.0
PyYAML                  6.0.1
pyzmq                   26.0.3
rasterio                1.4a3
requests                2.32.2
resnest                 0.0.6b20230402
rich                    13.7.1
ruamel.yaml             0.17.21
ruamel.yaml.clib        0.2.6
safetensors             0.4.3
sahi                    0.11.16
scikit-image            0.19.3
scikit-learn            1.5.0
scipy                   1.13.1
seaborn                 0.13.2
sentry-sdk              2.7.1
setproctitle            1.3.3
setuptools              69.5.1
shapely                 2.0.4
shellingham             1.5.4
six                     1.16.0
smart-open              7.0.4
smmap                   5.0.1
spacy                   3.7.5
spacy-legacy            3.0.12
spacy-loggers           1.0.5
srsly                   2.4.8
stack-data              0.6.3
tabulate                0.9.0
tensorboard             2.17.0
tensorboard-data-server 0.7.2
termcolor               2.4.0
terminaltables          3.1.10
text-unidecode          1.3
thinc                   8.2.5
threadpoolctl           3.5.0
tifffile                2024.6.18
timm                    1.0.7
tomli                   2.0.1
torch                   1.10.0+cu111
torchmetrics            1.4.0.post0
torchtext               0.11.0
torchvision             0.11.1+cu111
tornado                 6.4.1
tqdm                    4.66.4
traitlets               5.14.3
typer                   0.12.3
typing_extensions       4.12.2
tzdata                  2024.1
urllib3                 2.2.2
wandb                   0.17.3
wasabi                  1.1.3
wcwidth                 0.2.13
weasel                  0.4.1
webencodings            0.5.1
Werkzeug                3.0.3
wheel                   0.43.0
wrapt                   1.16.0
yacs                    0.1.8
yapf                    0.40.2
yarl                    1.9.4
yolov5-icevision        6.0.0
zipp                    3.19.2
zstandard               0.22.0
